% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConcepts.R
\name{getConcepts}
\alias{getConcepts}
\title{Get concepts from different fraily Indices}
\usage{
getConcepts(index)
}
\arguments{
\item{index}{character vector, one of "efi", "vafi", "efragicap", "hfrs"}
}
\value{
A tibble with columns labelled "category" and "concept_id"
}
\description{
Funtion to retrieve table of OMOP concept_ids and their labels
}
