% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop2efi.R
\name{getEligible}
\alias{getEligible}
\title{Get Eligible FI cohort}
\usage{
getEligible(con)
}
\arguments{
\item{con}{}
}
\value{
dataframe of elgible IDs
}
\description{
Get single column data frame of eligible IDs named person_id
}
