% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFI.R
\name{getFI}
\alias{getFI}
\title{Title}
\usage{
getFI(
  .data,
  weighted_fi = NA,
  person_id = "person_id",
  concept_id = "concept_id",
  concept_name = "concept_name",
  concept_start_date,
  .data_search,
  search_person_id,
  search_start_date,
  interval = 365,
  summary,
  group_var,
  rejoin = FALSE,
  ...
)
}
\arguments{
\item{.data}{data.frame; FI dataframe of all data (e.g., efi_dat.csv)}

\item{person_id}{chr; column name of person_id within .data_search dataframe}

\item{.data_search}{data.frame; dataframe of participant IDs, start, and end dates to be searched}

\item{interval}{int; number of days to search from start date. defaults to 365}

\item{summary}{log; whether to return summary data or results broken out by FI concept}

\item{group_var}{chr; whether to summarize by "person" (default) or "category"}

\item{rejoin}{log; if TRUE, will join results back to .data_search. defaults to FALSE}

\item{...}{; additional grouping variables for summarizing. Column name in .data_search (E.g., calculate by year.)}

\item{index}{chr; name of FI index (efi, vafi, hfrs)}

\item{start_date}{chr; column name of start date column within .data_search. YYYY-MM-DD}
}
\value{
A dataframe of FI indices or raw values
}
\description{
Title
}
