% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd2fi.R
\name{icd2fi}
\alias{icd2fi}
\title{OMOP to FI}
\usage{
icd2fi(
  con,
  index,
  schema = NULL,
  collect = FALSE,
  .data_search,
  search_person_id,
  search_start_date,
  search_end_date,
  keep_columns = NULL,
  unique_categories = FALSE,
  concept_location
)
}
\arguments{
\item{con}{a database connectino using dbConnect() or similar}

\item{index}{chr vector; a frailty index. One of "efi", "efragicap", "vafi", or "hfrs"}

\item{schema}{chr vector: a character vector of the schema holding the table. defaults to NULL (no schema)}

\item{collect}{log; should the query be collected at the end or kept as an SQL query? This must be TRUE for bigquery
because bigquery does not permit temporary tables.}

\item{.data_search}{sql; An SQL query from the same connection source with a column for person_id, start_date, and end_date.}

\item{search_person_id}{chr; The name of the person_id column}

\item{search_start_date}{chr/date; the name of the start_date column}

\item{search_end_date}{chr/date; the name of the end_date column}

\item{keep_columns}{chr vector: any additional columns to keep in the final dataframe}

\item{unique_categories}{log; Should the result return just the distinct FI category occurances in the interval
all concept_ids and concepts taht occcur in the interval for each person (much larger result)}

\item{concept_location}{pointer to a table that holds the concepts for the FI. Used when collect = FALSE.
I recommend pre-uploading the aouFI::fi_indices table to your schema in the database. If the database
is small enough, you can instead set collect to TRUE and concept_location to NA to use a local table from aouFI.}
}
\value{
dataframe with EFI occurences that can be summarized
}
\description{
Generates table of condition occurances using ICD codes from an OMOP database for
the EGRAGICAP, VAFIfrailty indices.
This requires a database connection using dbConnect() or similar R package
The connection should be able to access the following OMOP CDM tables
\itemize{
\item concept
\item condition_occurrence
\item observation
\item procedure_occurrence
\item device_exposure
}
}
\examples{
# For All of Us, set collect = TRUE because no temp tables
omop2fi(con = con,
        index = "efi",
        collect = TRUE,
        .data_search = demo2,
        search_person_id = "person_id",
        search_start_date = "survey_date",
        search_end_date = "end_date",
        keep_columns = c("dob", "sex_at_birth"),
        unique_categories = TRUE,
        concept_location = NA
)

# Otherwise, we can tell the function where to look for the concepts
omop2fi(con = con,
        index = "efi",
        collect = FALASE,
        .data_search = demo2,
        search_person_id = "person_id",
        search_start_date = "survey_date",
        search_end_date = "end_date",
        keep_columns = c("dob", "sex_at_birth"),
        unique_categories = TRUE,
        concept_location = tbl(con, inDatabaseSchema(my_schema, "efi")
)

}
