% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop2efi.R
\name{omop2efi}
\alias{omop2efi}
\title{OMOP to EFI}
\usage{
omop2efi(con, eligible)
}
\arguments{
\item{con}{}
}
\value{
dataframe with EFI occurences
}
\description{
Generates table of EFI-related condition occurances from an OMOP database.
}
\examples{
con <- dbConnect(
bigrquery::bigquery(),
    billing = Sys.getenv('GOOGLE_PROJECT'),
    project = prefix,
    dataset = release
    )
omop2efi(con = con, eligible = getEligible())

}
