% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop2fi.R
\name{omop2fi}
\alias{omop2fi}
\title{OMOP to FI}
\usage{
omop2fi(con, eligible, index)
}
\arguments{
\item{con}{a database connectino using dbConnect() or similar}

\item{eligible}{a dataframe or tibble with a single column of unique person_id values. Obtained
in the all of us databse using getEligible() which returns a single column dataframe with
participatns older than 50 and younger than 120. For other datasources, will need to be custom made.}

\item{index}{chr vector; a frailty index. One of "efi", "efragicap", "vafi", or "hfrs"}
}
\value{
dataframe with EFI occurences that can be summarized into an EFI using aouFI::getFI()
}
\description{
Generates table of condition occurances from an OMOP database for
the EFI, EGRAGICAP, VAFI, and HFRS frailty indices.
This requires a database connection using dbConnect() or similar R package
The connection should be able to access the following OMOP CDM tables
\itemize{
\item concept
\item condition_occurrence
\item observation
\item procedure_occurrence
\item device_exposure
}
}
\examples{
con <- dbConnect(
bigrquery::bigquery(),
    billing = Sys.getenv('GOOGLE_PROJECT'),
    project = prefix,
    dataset = release
    )
omop2fi(con = con, eligible = aouFI::getEligible(), index = "efi")

}
