% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop2hfrs.R
\name{omop2hfrs}
\alias{omop2hfrs}
\title{OMOP to HFRS}
\usage{
omop2hfrs(con, eligible)
}
\arguments{
\item{con}{a database connectino using dbConnect() or similar}

\item{eligible}{a dataframe or tibble with a single column of unique person_id values. Obtained
in the all of us databse using getEligible() which returns a single column dataframe with
participatns older than 50 and younger than 120.}
}
\value{
dataframe with EFI occurences that can be summarized into an EFI using aouFI::getFI()
}
\description{
Generates table of HFRS-related condition occurances from an OMOP database.
This requires a database connection using dbConnect() or similar R package
The connection should be able to access the following OMOP CDM tables
\itemize{
\item concept
\item condition_occurrence
\item observation
\item procedure_occurrence
\item device_exposure
}
}
\examples{
con <- dbConnect(
bigrquery::bigquery(),
    billing = Sys.getenv('GOOGLE_PROJECT'),
    project = prefix,
    dataset = release
    )
omop2efi(con = con, eligible = aouFI::getEligible())

}
