% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeFI.R
\name{summarizeFI}
\alias{summarizeFI}
\title{Title}
\usage{
summarizeFI(
  .data,
  weighted_fi = NA,
  person_id = "person_id",
  concept_id = "concept_id",
  category = "category",
  concept_start_date = "start_date",
  summary,
  group_var,
  rejoin = FALSE
)
}
\arguments{
\item{.data}{data.frame; FI dataframe of all data (e.g., efi_dat.csv)}

\item{person_id}{chr; column name of person_id within .data_search dataframe}

\item{summary}{log; whether to return summary data or results broken out by FI concept}

\item{group_var}{chr; whether to summarize by "person" (default) or "category"}

\item{rejoin}{log; if TRUE, will join results back to .data_search. defaults to FALSE}

\item{index}{chr; name of FI index (efi, vafi, hfrs)}

\item{.data_search}{data.frame; dataframe of participant IDs, start, and end dates to be searched}

\item{start_date}{chr; column name of start date column within .data_search. YYYY-MM-DD}

\item{interval}{int; number of days to search from start date. defaults to 365}
}
\value{
A dataframe of FI indices or raw values
}
\description{
Title
}
