% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConcepts.R
\name{getConcepts}
\alias{getConcepts}
\title{Get concepts from different fraily Indices}
\usage{
getConcepts(index)
}
\arguments{
\item{index}{character vector, one of "efi", "vafi", "efragicap", "hfrs"}
}
\value{
a dataframe of FI categories and related concepts OR ICD codes
}
\description{
Get concepts from different fraily Indices
}
