% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEligible.R
\name{getEligible}
\alias{getEligible}
\title{Get single column data frame of eligible IDs}
\usage{
getEligible(con)
}
\arguments{
\item{con}{}
}
\value{
dataframe of elgible IDs
}
\description{
Get single column data frame of eligible IDs
}
