% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEligible.R
\name{getEligible}
\alias{getEligible}
\title{Get single column data frame of eligible IDs}
\usage{
getEligible(con, lower = 50, upper = 120)
}
\arguments{
\item{con}{connection to AoU database}

\item{lower}{lower bound age limit, inclusive. default 50}

\item{upper}{upper bound age limit, inclusive. default 120}
}
\value{
sql query for elgible IDs
}
\description{
Get single column data frame of eligible IDs
}
