% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFI.R
\name{getFI}
\alias{getFI}
\title{Title}
\usage{
getFI(
  .data,
  index,
  .data_search,
  person_id,
  start_date,
  end_date,
  summary,
  group_var
)
}
\arguments{
\item{.data}{data.frame; FI dataframe of all data (e.g., efi_dat.csv)}

\item{index}{chr; name of FI index (efi, vafi, hfrs)}

\item{.data_search}{data.frame; dataframe of participant IDs, start, and end dates to be searched}

\item{person_id}{chr; column name of person_id within .data_search dataframe}

\item{start_date}{chr; column name of start date column within .data_search. YYYY-MM-DD}

\item{end_date}{chr; column name of end date column within .data_search. YYYY-MM-DD}

\item{summary}{log; whether to return summary data or results broken out by FI concept}

\item{group_var}{chr; whether to summarize by "person" (default) or "category"}
}
\value{
A dataframe of FI indices or raw values
}
\description{
Title
}
\examples{
\dontrun{
getFI(.data = efi_dat,
index = "efi",
.data_search = test_dat,
person_id = "person_id",
start_date = "start_date",
end_date = "end_date",
group_var = "person_id",
summary = TRUE)
}
}
